requirements.txt
# AMATS - Access Management and Asset Tracking System
# UTV Ghana Field Study Project

# Core
Django>=5.0,<6.0
djangorestframework>=3.14
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
Pillow>=10.0.0  # For image processing (QR codes)
python-dotenv>=1.0.0  # For environment variables

# Security & utilities
django-csp>=3.7
django-ratelimit>=4.0
django-cryptography>=1.0

# Realtime & async
channels>=4.0
channels-redis>=4.0
redis>=4.5

# Network scanning
python-nmap>=0.7.1
scapy>=2.5.0

# Testing & CI
pytest>=7.0
pytest-django>=4.5
coverage>=6.0

# Production
gunicorn>=20.1.0
psycopg2-binary>=2.9
celery>=5.2
django-celery-beat>=2.4
django-celery-results>=2.5
argon2-cffi>=23.0
